requirements.txt
aiohttp==3.5.4
aiopg==0.16.0
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
psycopg2-binary==2.7.7
PyYAML==5.1
SQLAlchemy==1.3.1
yarl==1.3.0
